Some simplifications and tests
