+++ tests/unit/test_datacontainer.py
@@ -67,28 +67,3 @@ def test_cc_add():
     assert c2.mapping() == [("a", "b"), ("b", "b"), ("c", "c")]
     assert c.columns == ["a", "b", "c"]
     assert c.mapping() == [("a", "a"), ("b", "b"), ("c", "c")]
-
-
-def test_schema_container():
-    example = SchemaContainer("example")
-    example.add("experiment", pd.DataFrame())
-    assert "experiment" in example.experiments
-
-    example.add("model", ("dummy_model", "dummy_cols"))
-    assert "model" in example.models
-
-    dummy_data_container = DataContainer(pd.DataFrame(), ColumnContainer(""))
-    example.add("table", dummy_data_container)
-    assert "table" in example.tables
-
-    example.add("funcs", lambda x: x + 42)
-    assert "funcs" in example.functions
-
-    example.drop("tables", "table")
-    assert "table" not in example.tables
-    example.drop("experiments", "experiment")
-    assert "experiment" not in example.experiments
-    example.drop("models", "model")
-    assert "model" not in example.models
-    example.drop("functions", "funcs")
-    assert "funcs" not in example.functions
